[new_feature]  将ssm中的blockid修改成uint64_t，在使用过程中出现的BUG see http://rb.corp.taobao.com/r/58836 将ssm中的blockid修改成uint64_t以及在使用过程中出现的BUG
git-svn-id: http://code.taobao.org/svn/tfs@2171 b722c2b5-6a69-4643-bcd1-1a75825fa998

+++ branches/tfs-tmp-trunk/src/tools/nameserver/repair_block_info.cpp
@@ -37,7 +37,7 @@ using namespace std;
 using namespace tfs::common;
 using namespace tfs::message;
 
-int get_block_ds_list(const uint64_t server_id, const uint32_t block_id, VUINT64& ds_list)
+int get_block_ds_list(const uint64_t server_id, const uint64_t block_id, VUINT64& ds_list)
 {
   int ret = TFS_SUCCESS;
   if (false == NewClientManager::get_instance().is_init())
